backend/requirements.txt
fastapi==0.115.12
fastapi-mail==1.5.0
uvicorn[standard]==0.34.2
pydantic[email]==2.11.4
pydantic-settings==2.10.1
motor==3.7.1
pymongo==4.13.2
python-jose[cryptography]==3.5.0
python-dotenv==1.0.0
openai==1.77.0
httpx==0.28.1
requests==2.32.3
requests-toolbelt==1.0.0
langchain-openai==0.3.16
python-multipart==0.0.6
reportlab==4.4.4
markdown2==2.5.4
pillow==12.0.0